More writing for the math and browser APIs
