+++ blog/2024-11-15-playing-with-fire/1-introduction/cameraGasket.ts
@@ -1,0 +1,10 @@
+export function camera(
+    size: number,
+    x: number,
+    y: number
+): [number, number] {
+    return [
+        Math.floor(x * size),
+        Math.floor(y * size)
+    ];
+}
