+++ blog/2024-11-15-playing-with-fire/1-introduction/chaosGame.js
@@ -30,5 +30,4 @@ function* chaosGame({width, height}) {
   yield img;
 }
 
-// Wiring so the code above displays properly
 render(<Gasket f={chaosGame}/>)
