+++ blog/2024-11-15-playing-with-fire/2-transforms/index.mdx
@@ -12,21 +12,14 @@ shapes and patterns that fractal flames are known for.
 <!-- truncate -->
 
 :::note
+This post uses a set of [reference parameters](../params.flame) to demonstrate the fractal flame algorithm.
+If you're interested in tweaking the parameters, or generating your own art, [Apophysis](https://sourceforge.net/projects/apophysis/)
+can load that file and you can try tweaking things yourself!
 
-This post uses a set of [reference parameters](../params.flame) to demonstrate a working
-implementation of the fractal flame algorithm. If you're interested in tweaking the parameters,
-or generating your own art, [Apophysis](https://sourceforge.net/projects/apophysis/)
-can load that file and gives full control over the image.
-
+This post covers section 3 of the Fractal Flame Algorithm paper
 :::
 
 ## Transforms and variations
-
-:::note
-
-This post covers section 3 of the Fractal Flame Algorithm paper
-
-:::
 
 import CodeBlock from '@theme/CodeBlock'
